requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
SQLAlchemy==2.0.31
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
Flask-Migrate==4.0.7